Use more specific lowercase mappings in input file
The output doesn't change with this. But it makes the input file more
consistent and potentially more useful.

Consistency: The input file already used a mixture of mappings to
uppercase as well as lowercase characters. It was just inconsistent
and confusing.

Why the output doesn't change: The file starts with mappings from
lowercase "a" to uppercase "A", and so on. From there on all following
mappings to either "A" or "a" are put in the same group "A". If you
want to create a different output file all you need to do is to either
remove the initial "a => A" mapping or change it to "a => a". This
patch makes this much easier.

Change-Id: Iac67a0d768386d0467ca9f271cbb1dfecb1f1418