Use much more specific mappings whenever possible
For example, instead of mapping "1FE Ǿ" directly to "O" we use an
indirect mapping to "D8 Ø", which maps to "O". This makes the data
more flexible to use and easier to discuss and change in later
patches.

Not you might find a lot of characters that look like I should have
updated them the same way, but didn't. So far this is intentional to
not make changes to the output. This is for later patches.

Change-Id: If64cf93f759f2f2a0f553eb275559a643050150d