Some more specific mappings without change to the output
Follow up for the comments in If64cf93.

The idea here is to map each character to the next one that is a
little bit more trivial, but still as close as possible. Usually
going up in the list of codepoints, but not always. For example,
"ᾦ" maps to "ᾠ", which maps to "ὠ", to "ω", and finally to "w".

Change-Id: If5c6a8ab5d31515a47a68a5ade3c1a201e2c6438