Prefer mapping lowercase to uppercase characters
This follows the very first rule in equivset.in: "Case folding.
Although letters of different cases are often visually distinct,
they an easily be confused by people who are familiar with the
alphabet. Two words with a different case may be read as the same
word. This is a popular technique for impersonation."

Basically: No other character can be closer than the one that is
exactly the same, just uppercase.

Note this patch intentionally doesn't change anything in the output.

Change-Id: I50ffeaa1e1f3879fa7e27b50cd1b318f605e0a42