Revert "Fix inconsistent mapping of Greek η to "n" and sometimes "h""
This reverts commit 8ad4c836987fd04208421990fcb8add9afff5635.

Reason for revert: That Greek character is not an "n". It is the small
version of the Greek letter "Η" (Eta).

The very first rule in equivset.in says: "Case folding. Although letters
of different cases are often visually distinct, they can easily be
confused by people who are familiar with the alphabet. Two words with a
different case may be read as the same word. This is a popular technique
for impersonation."

This is important for people that can read Greek. For example, the Greek
name "Heros" is "ἥρως" in lower and "ἭΡΩΣ" in uppercase and should be
detected as similar. But it currently maps to "NPΩC" and "HPΩΣ". This
doesn't make sense.

This was really a mistake I made in 8ad4c836. Let's just undo it.

Change-Id: I89c204af167e590ab8d738837b7b7cff721689d5