More specific Greek mappings without change to the output
This is a direct follow up for If5c6a8a. While the mappings from that
patch are not wrong, I did some of them without consulting the actual
Unicode sources. The source for the new mappings is purely
https://www.unicode.org/charts/PDF/U1F00.pdf

Note how this doesn't change anything in the compiled files. It's
purely cosmetic, to make the input file as precise as possible.

Change-Id: I33f37e8e427463ad0db3a028696033905a1e6a1e