Fix mapping of Latin/Cyrillic capital letter Iota
The ASCII letters I (the uppercase i) and l (the lowercase L) are
always very easy to confuse, depending on the font. Still the
decision was made to not merge the groups I and L.

This creates the challenge that any letter can either be mapped to
I or L, but never both. This decision needs to be made, but is often
quite arbitrary.

While it's possible to argue that an Iota also looks like a lowercase
l, it's equally valid to map it to I. If not more valid. It's called
"Iota", after all.

Notice how the lowercase Iota was already mapped to "i". Let's fix
this inconsistency.

The Cyrillic Iota was in a separate group before. But since it's
pretty much the same letter I merge it into the same group.

Change-Id: Id5ed541e8cafa9bccefdf9ba370fd466bb885536