Fix mapping of Cyrillic letter Be
I'm not sure how this happened. This letter is named "Be" and
clearly looks like one, doesn't it?

While it might be possible to argue that it also looks a bit like
the digit 6 (a really weird one, though), doing so puts it into the
group G. It really doesn't belong there.

The Unicode chart links the Cyrillic Б to the Latin ƃ (small letter
b with topbar). That's already in the group B, obviously.

Change-Id: I9f238b6f4c75f9b6243241e88291e1e1fa2cdf0f