Fix inconsistent mapping of Cherokee letter Tsu
I believe this was just a mistake.

Cherokee traditionally doesn't have capitalization. That's why the
Unicode group for the small letters is so far away, and why the
letters look exactly the same, just smaller.

Because of this it really doesn't make sense to use different
mappings for the two versions.

While I can see how "stroke down and bending to the left" can be
used to describe both the shape of the Cherokee Ꮷ and the uppercase
J, the bend in Ꮷ forms a closed circle, but J does not.

Mapping it to d makes much more sense, in my opinion. Note how the
small version was already mapped to d.

Change-Id: I0cf3e0940caa42632310e07c830f225823b2dfbc