Map Cyrillic ҽ (Abkhasian Che) to lowercase e
The letter is described as "Cyrillic small letter Abkhasian Che".
It might be valid to map it to the equally named "Cyrillic small
letter Che", but that looks totally different: ч. Instead I
suggest to follow the "visually similar" rule at the top of
equivset.in and map it to e, because it clearly looks like one.

Change-Id: Iba84d0c9a510edb59b015436eb9d3874aa32da39