Map old Coptic letters to the Greek range
Note how the order is almost the exact same in both ranges, starting
with 2C80 Ⲁ (ALFA) = 391 Α (ALPHA), and so on.

I'm intentionally turning some mappings around from lower→uppercase
to upper→lowercase when the old Coptic letter looks much more like
the lowercase Greek letter. This doesn't make a difference in the
output.

Change-Id: I39938f7208503cad769d5a170b591dd52d26f56e