Map identical Hiragana and Katakana letters together
Please have a look at the Unicode charts to understand this better:
https://www.unicode.org/charts/PDF/U3040.pdf
https://www.unicode.org/charts/PDF/U30A0.pdf
Not only do these letters have the exact same names, they are in the
exact same order in the two charts. The idea is that users who know
these languages know what these letters mean and how they are spelled
and use them interchangeably.

On top of that some look very similar (e.g. 304B か and 30AB カ) or
even the exact same (e.g. 3078 へ and 30D8 ヘ, which was already
mapped before).

Change-Id: I745d72615faa2c9e21d337a91ae59f642f98464c