Detect more possible mistakes in the generator
1. I made a mistake in I39938f7 and mapped a character to itself.
This should be detected as an error.

This patch also fixes the mistake.

2. It's possible and even useful to create circles with two or more
elements, e.g. map a→A and A→a. It doesn't matter much which of the
characters represents the group in the end, as long as they all end
in the same group. This works already, with one exception: Such a
group is entirely dropped from the .txt file.

I added a unit test for this edge-case. Luckily we currently don't
have such a problem in the .in file. But I run into this when I
prepared my recent submissions.

Change-Id: I72554e6af6d2f48c0d1458851f0b664998a2c3c6