Map some very obvious Hangul Jamo look-alikes to ASCII
I think there are some very obvious look-alikes in this range
(especially the full-height versions) that can very easily be used
as fake Latin/ASCII characters. This follows rule #2 at the top of
equivset.in.

There are a few more, e.g. ㅠ looks like Pi, as well as ᄅ and ᄀ
that look like numbers. But this is probably much less relevant.

Bug: T357855
Change-Id: I2ccb8dfe3a6944a1568beb827267a9b3fc190654